% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildnex.R
\name{buildnex}
\alias{buildnex}
\title{Read a nexus data file}
\usage{
buildnex(file, charlabels = NULL, charnums = NULL,
  statelabels = NULL, taxlabels = NULL, filename = NULL,
  ntax = NULL, nchar = NULL, first = NULL, last = NULL,
  missing = "?", gap = "-")
}
\arguments{
\item{file}{(required) path to either a PDF file or a TXT file}

\item{charlabels}{optional path to text file with character statements}

\item{charnums}{optional vector with character numbers}

\item{statelabels}{optional character vector with state labels for characters}

\item{taxlabels}{optional vector with names of taxa}

\item{filename}{vector with names of files associated with characters}

\item{ntax}{number of taxa (needed for PDF file reading)}

\item{nchar}{number of characters (needed for PDF file reading)}

\item{first}{first page for data matrix in a PDF file}

\item{last}{last page for data matrix in a PDF file}

\item{missing}{character representing missing data}

\item{gap}{character representing inapplicable/incomporable data}
}
\value{
an object of class \code{nex} for use in further \code{nexustools} functions
}
\description{
A function that reads data stored in nexus files
NOTE: it is necessary to first install poppler with, e.g., `brew install poppler`
}
\examples{
\dontrun{
tmp <- buildnex('/Users/chadeliason/github/nexustools/example/Bertelli_Chiappe_2005.pdf', ntax = 34,
nchar = 63, first = 22, last = 22)
plot(tmp, legend.pos = 'top')
}

}
\author{
Chad Eliason \email{chad_eliason@utexas.edu}
}
