% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_comments.R
\name{capture_comments}
\alias{capture_comments}
\title{Function to find comments in text file}
\usage{
capture_comments(file)
}
\arguments{
\item{file}{name of file}
}
\description{
How characters are characterized
DD = duplicate
KK = keep
KKM = keep and modify wording
XX = cut
}
