% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleantext.R
\name{cleantext}
\alias{cleantext}
\title{Function to remove spaces, punctuation from phenome terms}
\usage{
cleantext(x, fast = TRUE, latin = TRUE, cuts = TRUE,
  comments = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{fast}{whether to use fast version (for duplicated analysis) or slow version (for printing characters)}

\item{latin}{whether to use Latin token algorithm (see Schinke 1996)}

\item{cuts}{whether to cut terms from a list}

\item{comments}{logical whether to remove test in square brackets}
}
\description{
Given a character vector, remove comments, superfluous characters, extra spaces
and non-anatomical terms
}
\examples{
\dontrun{
cleantext("dorsalmost part of the processus ligamentus cranii", fast=FALSE, cuts=FALSE)
}

}
\author{
Chad M. Eliason
}
