% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Write a nexus data object to a file}
\usage{
collapse(x, map, method = c("retain", "merge"))
}
\arguments{
\item{x}{(required) a `nex` object}

\item{map}{a list specifying equivalent taxa to map from and to (e.g., `c('Tinamus' = 'Tinamus_major')` will map all characters from Tinamus to Tinamus_major)}

\item{method}{whether to convert characters scored in both taxa to polymorphisms (`merge`) or retain original characters (default)}
}
\value{
an object of class \code{nex} for use in further \code{phenotools} functions
}
\description{
Function to collapse taxa and merge data in taxon 1 with other taxa
}
\author{
Chad Eliason \email{celiason@fieldmuseum.org}
}
