% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_kuhr.R
\name{curate_kuhr}
\alias{curate_kuhr}
\title{Curate kuhr data}
\usage{
curate_kuhr(
  diagnoses,
  code_system = "ICPC",
  primary_care_only = TRUE,
  recursive = TRUE,
  exclusions = NULL,
  group_all = FALSE,
  dx_groupname = NULL,
  dx_owners = c("child"),
  dx_range_limits = NULL,
  kuhr_full = NULL,
 
    moba_data_root_dir = "//ess01/P471/data/durable/data/MoBaPhenoData/PDB2306_MoBa_V12/SPSS/",
  kuhr_data_root_dir = "//ess01/P471/data/durable/data/KUHR/KUHR_current/DSV files/",
  kuhr_filenames_override = NULL,
 
    kuhr_linkage_file_root_dir = "//ess01/P471/data/durable/data/Linkage_files/KUHR_link/",
  kuhr_linkage_filename = "PDB2306_kobling_KUHR_mor_far_barn.sav",
  PDB = "2306",
  moba_data_version = 12,
  moba_filepaths = NULL,
  ...
)
}
\arguments{
\item{diagnoses}{Diagnostic codes (see \code{available_variables(source=\"npr\")})}

\item{code_system}{can be "ICPC" (the default) or "ICD10"}

\item{primary_care_only}{the default is to use only primary care data sources
from kuhr (so this option is set to TRUE); if you have other data sources in
your kuhr datafile, set this to FALSE}

\item{recursive}{also include sub-codes of codes you have provided in the
diagnoses field? Defaults to TRUE (only used with code_system = "ICD10")}

\item{exclusions}{codes to exclude - always non-recursive (i.e., only the
specific codes you provide will be excluded from individual-level summaries, not sub-codes)
(only used with code_system = "ICD10")}

\item{group_all}{If TRUE, will treat all your codes as a single diagnosis
for the purpose of summarising; defaults to FALSE}

\item{dx_groupname}{If group_all is TRUE, this string is used to label the
vareiables pertaining to the grouped diagnoses}

\item{dx_owners}{Whose diagnoses do you want to count? Dataset is returned as
one row per pregnancy, but you can get the diagnoses relevant to any
combination of "child", "father", and "mother"; the default is "child"}

\item{dx_range_limits}{If you want only to extract diagnoses from a particular
range of available years, provide limits here (e.g., "c(2008,2010)" will retrieve
diagnoses from 2008, 2009, and 2010 only). Defaults to NULL - i.e., extract
from all available data}

\item{kuhr_full}{if you already have the KUHR data loaded in memory, you can
save some time by providing the name of the R object here; note
that there are specific requirements for the structure of this file (see
Details) below.}

\item{moba_data_root_dir}{Where is the raw MoBa phenotypic data? (default
is for p471)}

\item{kuhr_data_root_dir}{Where is the raw NPR phenotypic data? (default
is for p471)}

\item{kuhr_filenames_override}{If the KUHR filenames are not YEAR_Data_KUHR.csv,
what are they?}

\item{kuhr_linkage_file_root_dir}{Where are the linkage files? (default
is for p471)}

\item{kuhr_linkage_filename}{What is the name of the linkage file? (default
is for p471)}

\item{PDB}{What is the PDB code for your TSD project? (default
is for p471)}

\item{moba_data_version}{What version is the MoBa data you are linking to? Defaults to
12}

\item{moba_filepaths}{Allows curate_dataset to pass on an amended list of filepaths}
}
\description{
\code{curate_kuhr} is called by \code{curate_dataset}
when kuhr variables are requested
}
\details{
Detailed description...
}
