% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_moba_scales.R
\name{curate_moba_scales}
\alias{curate_moba_scales}
\title{Curate moba scales}
\usage{
curate_moba_scales(moba_scale_vars, moba_scale_vars_long, moba_data,
  completion_threshold = 0.5, return_items = FALSE,
  consistent_items = FALSE, transformations = NULL)
}
\arguments{
\item{moba_scale_vars}{Df created in \code{curate_dataset}}

\item{moba_scale_vars_long}{Long format Df created in \code{curate_dataset}}

\item{moba_data}{Df created in \code{curate_dataset}}

\item{completion_threshold}{What proportion of scale items need
to be present for a scale score to be computed?}

\item{return_items}{Output item-level data?}

\item{consistent_items}{Only use wave-to-wave consistent items
for variables that are measured longitudinally?}

\item{transformations}{Not yet implemented}
}
\description{
\code{curate_moba_scales} is is called by \code{curate_dataset}
when moba scale variables are requested
}
\details{
Detailed description...
}
