% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_npr.R
\name{curate_npr}
\alias{curate_npr}
\title{Curate NPR data}
\usage{
curate_npr(
  diagnoses,
  recursive = TRUE,
  exclusions = NULL,
  group_all = FALSE,
  dx_groupname = NULL,
  dx_owners = c("child"),
  npr_full = NULL,
  npr_preprocessed = NULL,
 
    moba_data_root_dir = "//tsd-evs/p471/data/durable/data/MoBaPhenoData/PDB2306_MoBa_V12/SPSS/",
  npr_data_root_dir = "//tsd-evs/p471/data/durable/data/NPR/processed/",
  npr_filename = "18_34161_NPR.sav",
  linkage_file_root_dir = "//tsd-evs/p471/data/durable/data/Linkage files/",
  npr_linkage_filename = "PDB2306_kobling_NPR_mor_far_barn.sav",
  PDB = "2306",
  moba_data_version = 12
)
}
\arguments{
\item{diagnoses}{ICD codes (see \code{available_variables(source=\"npr\")})}

\item{recursive}{also include sub-codes of codes you have provided in the
diagnoses field? Defaults to TRUE}

\item{exclusions}{codes to exclude - always non-recursive (i.e., only the
specific codes you provide will be excluded from individual-level summaries, not sub-codes)}

\item{group_all}{If TRUE, will treat all your codes as a single diagnosis
for the purpose of summarising; defaults to FALSE}

\item{dx_groupname}{If group_all is TRUE, this string is used to label the
vareiables pertaining to the grouped diagnoses}

\item{dx_owners}{Whose diagnoses do you want to count? Dataset is returned as
one row per pregnancy, but you can get the diagnoses relevant to any
combination of "child", "father", and "mother"; the default is "child"}

\item{npr_full}{if you already have the NPR data loaded in memory, you can
save some time by providing the name of the R object here; note
that there are specific requirements for the structure of this file (see
Details) below.}

\item{npr_preprocessed}{if you have already got a version of the NPR data
linked with MoBa IDs loaded in memory, supply the R object name here; note
that there are specific requirements for the structure of this file (see
Details) below.}

\item{moba_data_root_dir}{Where is the raw MoBa phenotypic data? (default
is for p471)}

\item{npr_data_root_dir}{Where is the raw NPR phenotypic data? (default
is for p471)}

\item{npr_filename}{What is the name of the NPR file? (default
is for p471)}

\item{linkage_file_root_dir}{Where are the linkage files? (default
is for p471)}

\item{npr_linkage_filename}{What is the name of the linkage file? (default
is for p471)}

\item{PDB}{What is the PDB code for your TSD project? (default
is for p471)}

\item{moba_data_version}{What version is the MoBa data you are linking to? Defaults to
12}
}
\description{
\code{curate_npr} is called by \code{curate_dataset}
when npr variables are requested
}
\details{
Detailed description...
}
