% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter duplicated characters in a nexus file}
\usage{
filter(x, dups = NULL)
}
\arguments{
\item{x}{(required) a nexus input file}

\item{dups}{an optional matrix or list specifying which character are duplicates (e.g., dups = list('1733' = c(1741,1745,1755)))}
}
\description{
Given a list of specified duplicates, drop and/or merge characters
}
\examples{
\dontrun{
dat <- read.nex("/Users/chadeliason/Dropbox/phenome dataset/data/2015-09-02/original/final_reordered.nex")
drops <- grep("livezey_2006", dat$charlab)
tmp <- capture_comments(file = "~/Dropbox/phenome dataset/data/2015-09-02/modified/final_reorderedJAC+CME.txt")
ss <- tmp$markup
x <- dat[, ss]
dups <- tmp$dups
xx <- filter(x, dups=tmp$dups)
sum(is.na(x$data))/length(x$data)  # 86.5\% missing data
sum(is.na(xx$data))/length(xx$data)  # 86.3\% missing data
}

}
