% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.nex.R
\name{findgroups}
\alias{findgroups}
\title{Sub-function to get groups of duplicated characters based on input map (char1 in column 1, char2 in column 2)}
\usage{
findgroups(x)
}
\arguments{
\item{x}{a nex object}
}
\description{
Sub-function to get groups of duplicated characters based on input map (char1 in column 1, char2 in column 2)
}
\author{
Chad Eliason \email{celiason@fieldmuseum.org}
}
