% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ontology.R
\name{generate_ontology}
\alias{generate_ontology}
\title{Generate trait ontology from a nexus file}
\usage{
generate_ontology(x)
}
\arguments{
\item{x}{a nexus file}
}
\description{
Comma separate terms are used to define links among components of a nexus
character list
}
