% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcomments.R
\name{getcomments}
\alias{getcomments}
\title{Function to find comments in text file}
\usage{
getcomments(file)
}
\arguments{
\item{file}{name of text file}
}
\description{
This function reads in a character list that have been annotated
as to whether characters are duplicates (see Details)
}
\details{
Text file contains list of characters with flags indicating whether
characters are duplicates (DD), should be cut (XX), good characters that
should be kept (KK), or characters that should be kept but modified (KKM).
Comments in curly braces indicate which other characters a focal character
duplicates (e.g., "{duplicates 125, 250}").
}
