% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nex.R
\name{head.nex}
\alias{head.nex}
\title{Look at top of a `nex` object}
\usage{
\method{head}{nex}(x, width = 6, ...)
}
\arguments{
\item{x}{nex object}

\item{width}{width of character string to display}

\item{...}{optional arguments}
}
\description{
Look at top of a `nex` object
}
