% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_project.R
\name{initialise_project}
\alias{initialise_project}
\title{Initialise project}
\usage{
initialise_project(
  path,
  template_filepath = "//tsd-evs/p471/data/durable/common/new_project_template/."
)
}
\arguments{
\item{path}{Where is the top-level directory for the new
project?}

\item{template_filepath}{Where is the template new project directory?
(Defaults to correct location for p471)}
}
\description{
\code{initialise_project} populates a specified directory
with the file structure required to make full use of phenotools
}
\details{
Detailed description...
}
