% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{Initialize project}
\usage{
initialize_project(
  path,
  template_filepath = "//ess01/P471/data/durable/common/new_project_template/."
)
}
\arguments{
\item{path}{Where is the top-level directory for the new
project?}

\item{template_filepath}{Where is the template new project directory?
(Defaults to correct location for p471)}
}
\description{
Alias of \code{initialise_project}
}
\details{
See documentation for \code{initialise_project}
}
