% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.nex.R
\name{merge.nex}
\alias{merge.nex}
\title{Merge a list of nexus files}
\usage{
\method{merge}{nex}(x, taxa = NULL)
}
\arguments{
\item{x}{(required) a list of nexus files to merge}

\item{taxa}{a character vector of names of taxa to include in concatenated nexus file (optional)}
}
\value{
an object of class \code{nex} for use in further \code{nexustools} functions
}
\description{
a function that combines numerous nexus datasets
}
\examples{
\dontrun{
x <- read.nex('example/toy1.nex')
y <- read.nex('example/toy2.nex')
xy <- merge.nex(list(x, y), taxa = c('species1', 'species2'))
xy
}
}
\author{
Chad Eliason \email{chad_eliason@utexas.edu}
}
