% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moba}
\alias{moba}
\title{Metadata for moba variables used in dataset curation}
\format{A data frame
\describe{
  \item{var_name}{variable level idenitfier used in
  phenotools package functions}
  ...
}}
\source{
make_raw_pkg_data.R in data-raw
}
\usage{
moba
}
\description{
A dataset containing details of each of the variables in
moba that phenotools can be used to make
}
\keyword{datasets}
