% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moba_bmi.R
\name{moba_bmi}
\alias{moba_bmi}
\title{BMI variable creation helper function}
\usage{
moba_bmi(moba_other_vars, moba_other_vars_long, moba_data)
}
\arguments{
\item{moba_other_vars}{Df made in \code{curate_dataset}}

\item{moba_other_vars_long}{Long form of Df made in \code{curate_dataset}}

\item{moba_data}{Df made in \code{curate_dataset}}
}
\description{
\code{moba_bmi} is a helper function
called by \code{curate_dataset} to code up BMI variables
}
\details{
Detailed description...
}
