% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moba_helpers.R
\name{moba_helpers}
\alias{moba_helpers}
\alias{single_item}
\alias{moba_mchat}
\alias{moba_bmi}
\alias{moba_lthmdd}
\title{Helper functions for creating MoBa variables with non-standard coding}
\usage{
single_item(
  moba_other_vars,
  moba_other_vars_long,
  moba_data,
  return_items,
  completion_threshold
)

moba_mchat(
  moba_other_vars,
  moba_other_vars_long,
  moba_data,
  return_items,
  completion_threshold
)

moba_bmi(
  moba_other_vars,
  moba_other_vars_long,
  moba_data,
  return_items,
  completion_threshold
)

moba_lthmdd(
  moba_other_vars,
  moba_other_vars_long,
  moba_data,
  return_items,
  completion_threshold
)
}
\arguments{
\item{moba_other_vars}{Df made in \code{curate_dataset}}

\item{moba_other_vars_long}{Long form of Df made in \code{curate_dataset}}

\item{moba_data}{Df made in \code{curate_dataset}}

\item{return_items}{Whether to return an item level dataset}

\item{completion_threshold}{What proportion of non-missing items in scale}
}
\description{
Some MoBa variables require bespoke coding. Sometimes this is because
of the way responses are coded or because of differences
between questionnaire versions. The helper functions within this file allow
\code{curate_dataset} to make these variables and include them in
datasets alongside standard scale variables and un-processed item retrievals.

NB: this is different to retrieval of single MoBa items by item code, as
those variables are returned without processing. Here we are dealing with
derived variables such as BMI, or scales where the standard coding in
\code{curate_moba_scales} isn't capable of handling some idiosyncrasies
}
\section{Non-standard MoBa variable helpers}{


The MoBa helper functions are listed below along with a short description
of the reason why a helper is needed. More info may be available in the notes
for a given variables (see notes column in \code{phenotools::moba} or get a
dataset report for these).

* `single_item()`: items that do not belong to a scale are retrieved by code using this function
* `moba_bmi()``: BMI variables must be derived based on weight and height; outliers are removed as part of this process
* `moba_mchat()`: some (but not all) M-CHAT variables have item codes that differ between questionnaire versions; a helper is needed to handle this
}

