% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.nex.R
\name{na.omit.nex}
\alias{na.omit.nex}
\title{Remove NAs from nexus files}
\usage{
\method{na.omit}{nex}(x, ...)
}
\arguments{
\item{x}{nex object}

\item{...}{other arguments}
}
\description{
Remove NAs from nexus files
}
