% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_curated_kuhr.R
\name{pivot_curated_kuhr}
\alias{pivot_curated_kuhr}
\title{Pivot curated -kuhr to long format to work with multiple occasions per individual}
\usage{
pivot_curated_kuhr(curated_kuhr, large_file = FALSE)
}
\arguments{
\item{curated_kuhr}{a dataframe produced by curate_kuhr/curate_dataset
with some all_dx variables in it}

\item{large_file}{do you have the RAM available to allow for a large file
(more than two sets of diagnosis/grouping variables) to be pivoted? Defaults
to FALSE (see Details below for suggestions if not)}
}
\description{
\code{pivot_curated_kuhr}  can be used on a dataframe containing kuhr variables
made by \code{curate_kuhr} ( either directly or, more commonly via \code{curate_kuhr})
to pivot the all_dx variables back to long format - i.e., taking the data back
to one row per healthcare contact
}
\details{
Detailed description...
}
