% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_curated_npr.R
\name{pivot_curated_npr}
\alias{pivot_curated_npr}
\title{Pivot curated NPR to long format to work with multiple occasions per individual}
\usage{
pivot_curated_npr(curated_npr, large_file = FALSE)
}
\arguments{
\item{curated_npr}{a dataframe produced by curate_npr/curate_dataset
with some all_dx variables in it}

\item{large_file}{do you have the RAM available to allow for a large file
(more than two sets of diagnosis/grouping variables) to be pivoted? Defaults
to FALSE (see Details below for suggestions if not)}
}
\description{
\code{pivot_curated_npr}  can be used on a dataframe containing NPR variables
made by \code{curate_npr} ( either directly or, more commonly via \code{curate_npr})
to pivot the all_dx variables back to long format - i.e., taking the data back
to one row per healthcare contact
}
\details{
Detailed description...
}
