% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload_kuhr.R
\name{preload_kuhr}
\alias{preload_kuhr}
\title{Pre-load KUHR data}
\usage{
preload_kuhr(
  kuhr_data_root_dir = "//tsd-evs/p471/data/durable/data/KUHR/processed/",
  kuhr_filenames_override = NULL,
  primary_care_only = "TRUE"
)
}
\arguments{
\item{kuhr_data_root_dir}{Where is the raw KUHR phenotypic data? (default
is for p471)}

\item{kuhr_filenames_override}{What is the name of the KUHR file? (default
is for p471)}

\item{primary_care_only}{can only be TRUE at present}
}
\description{
\code{preload_nor} returns a dataframe with the full KUHR dataset, useful if
you want to run curate_dataset multiple times for different KUHR variables
without waiting for the KUHR load each time; use in conjunction with the
'kuhr_full' option in curate_kuhr (passed on by curate_dataset).
}
\details{
Detailed description...
}
