% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload_npr.R
\name{preload_npr}
\alias{preload_npr}
\title{Pre-load NPR data}
\usage{
preload_npr(
  npr_data_root_dir = "//tsd-evs/p471/data/durable/data/NPR/processed/",
  npr_filename = "18_34161_NPR.sav"
)
}
\arguments{
\item{npr_data_root_dir}{Where is the raw NPR phenotypic data? (default
is for p471)}

\item{npr_filename}{What is the name of the NPR file? (default
is for p471)}
}
\description{
\code{preload_nor} returns a dataframe with the full NPR dataset, useful if
you want to run curate_dataset multiple times for different NPR variables
without waiting for the NPR load each time; use in conjunction with the
'npr_full' option in curate_npr (passed on by curate_dataset).
}
\details{
Detailed description...
}
