% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nex.R
\name{print.nex}
\alias{print.nex}
\title{Utility function to print metadata for nexus file}
\usage{
\method{print}{nex}(x, ...)
}
\arguments{
\item{x}{nex class object}

\item{...}{other arguments (not used)}
}
\description{
Utility function to print metadata for nexus file
}
