% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printneat.R
\name{printneat}
\alias{printneat}
\title{Print character names and states}
\usage{
printneat(x, y)
}
\arguments{
\item{x}{nexus object}

\item{y}{numeric vectors of character numbers to print}
}
\description{
Print character names and states
}
