% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_scale_items.R
\name{query_scale_items}
\alias{query_scale_items}
\title{Quickly query the items in a scale, and how phenotools combines them}
\usage{
query_scale_items(scales)
}
\arguments{
\item{scales}{a vector of phenotools variable names to search for}
}
\description{
\code{query_scale_items} is a helper to find items for a given phentools scale,
their raw codings, and how phenotools combines them to make the scale
}
\details{
Detailed description..
}
