% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dataset.R
\name{sim_data}
\alias{sim_data}
\title{Simulate dataset}
\usage{
sim_data(dataset, ...)
}
\arguments{
\item{dataset}{A data.frame with variables to be simulated}
}
\description{
\code{sim_dataset} returns a dataframe with simulated
data suitable for sharing with analysis code
}
\details{
Detailed description...
}
