% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_item.R
\name{single_item}
\alias{single_item}
\title{Single item variable creation helper function}
\usage{
single_item(moba_other_vars, moba_other_vars_long, moba_data)
}
\arguments{
\item{moba_other_vars}{Df made in \code{curate_dataset}}

\item{moba_other_vars_long}{Long form of Df made in \code{curate_dataset}}

\item{moba_data}{Df made in \code{curate_dataset}}
}
\description{
\code{single_item} is a helper function
called by \code{curate_dataset} to return single item variables
}
\details{
Detailed description...
}
