% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.nex.R
\name{sort.nex}
\alias{sort.nex}
\title{Sort a nexus file}
\usage{
\method{sort}{nex}(x, decreasing = FALSE, by = c("taxlabels", "charlabels"), ...)
}
\arguments{
\item{x}{nexus object}

\item{decreasing}{whether to sort in decreasing order}

\item{by}{sorting by}

\item{...}{other arguments passed to sort}
}
\description{
Sort a nexus file
}
\examples{
\dontrun{
data(twig)
twig.sorted <- sort(twig, by="taxlabels")
plot(twig.sorted)
}

}
