% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.nex.R
\name{subset.nex}
\alias{subset.nex}
\alias{[.nex}
\title{subset a nexus file}
\usage{
\method{subset}{nex}(x, condition)

\method{[}{nex}(x, i, j = NULL, drop = FALSE, ...)
}
\arguments{
\item{x}{nexus file}

\item{condition}{subset condition}

\item{i}{rows (taxa) to keep}

\item{j}{columns (characters) to keep}

\item{drop}{unused}

\item{...}{other arguments (not used)}
}
\description{
subset a nexus file

subset a nexus file
}
