% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nex.R
\name{summary.nex}
\alias{summary.nex}
\title{Summarize data in a `nex` object}
\usage{
\method{summary}{nex}(object, ...)
}
\arguments{
\item{object}{`nex` class object}

\item{...}{other arguments (not used)}
}
\description{
Summarize data in a `nex` object
}
