% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.nex.R
\name{tail.nex}
\alias{tail.nex}
\title{Look at bottom of a nexus file}
\usage{
\method{tail}{nex}(x, ...)
}
\arguments{
\item{x}{nex object}

\item{...}{optional arguments}
}
\description{
Look at bottom of a nexus file
}
