% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.nex.R
\name{update.nex}
\alias{update.nex}
\title{Small function to update a duplicated nex object}
\usage{
\method{update}{nex}(object, cutoff = Inf, ...)
}
\arguments{
\item{object}{nex object}

\item{cutoff}{cutoff to filter duplicates/overlapping characters by}

\item{...}{additional arguments (unused)}
}
\description{
Changing cutoff will affect the characters identified as duplicates for later
visualization (e.g., with `printout` or `duptree` functions)
}
\author{
Chad Eliason \email{celiason@fieldmuseum.org}
}
