% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_variables.R
\name{available_variables}
\alias{available_variables}
\title{Available variables}
\usage{
available_variables(source = c("moba", "npr", "kuhr"))
}
\arguments{
\item{source}{Currently any combination of "moba", "kuhr", "npr}
}
\value{
a data.frame with variable names and basic info
}
\description{
\code{available_variables} returns a dataframe with names and basic info
of all variables available for use via phenotools
}
\details{
Detailed description...
}
