% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_variables.R
\name{available_variables}
\alias{available_variables}
\title{Available variables}
\usage{
available_variables(source = "moba")
}
\arguments{
\item{source}{Currently "moba", "kuhr", "npr}
}
\description{
\code{available_variables} returns a dataframe with names and basic info
of all variables available for use via phenotools
}
\details{
Detailed description...
}
