% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_dataset.R
\name{curate_dataset}
\alias{curate_dataset}
\title{Curate dataset}
\usage{
curate_dataset(
  variables_required,
 
    moba_data_root_dir = "//ess01/P471/data/durable/data/MoBaPhenoData/PDB2306_MoBa_V12/SPSS/",
  PDB = "2306",
  moba_data_version = 12,
  completion_threshold = 0.5,
  return_items = FALSE,
  consistent_items = FALSE,
  transformations = NULL,
  log = NULL,
  out_format = "list",
  override_filenames = NULL,
  ...
)
}
\arguments{
\item{variables_required}{What variables are required? See
\code{available_variables} for a valid inputs}

\item{moba_data_root_dir}{Where is the raw MoBa phenotypic data?}

\item{PDB}{What is the PDB code for your TSD project?}

\item{moba_data_version}{Which version of MoBa phenotypic data are you using?
Defaults to 12; 11 is also a viable option for some scales}

\item{completion_threshold}{What proportion of scale items need
to be present for a scale score to be computed?}

\item{return_items}{Output item-level data?}

\item{consistent_items}{Only use wave-to-wave consistent items
for variables that are measured longitudinally?}

\item{transformations}{Not yet implemented}

\item{log}{filepath for code_preparation log - not yet implemented}

\item{out_format}{should output be a "list" or "merged_df"; defaults to "list"}

\item{override_filenames}{questionnaire filenames are built by combining your
inputs on PDB, moba_data_version, and the standard codes for each questionnaire
(see unique(available_variables("moba")$questionnaire) + "SV_INFO" and "MBRN")
to give the format: "PDB>PDB<_>code<_v>moba_data_version<.sav"). If you need to
override this because the filenames have been altered in your TSD project, add
 a vector of strings here, e.g., 'override_filenames = c("Q1 = new_Q1_filename.sav").
 The SV_INFO and MBRN files are accessed implicitly for every curate call, so
 you will need to provide filenames for these (as well as the questionnaire files
 from which your variables are to be sourced) if they differ from the expected format
 is a reminder that these sources}

\item{...}{arguments to pass to internal functions, see ?curate_moba_scales
and ?curate_npr}
}
\description{
\code{curate_dataset} is the main workhorse of the data preparation
component of phenotools, producing analysis-ready datasets with
specified variables from MoBa and other registry sources
}
\details{
This function fulfills the main purpose of the phenotools package by taking
details of variables you want in your analytic dataset, retrieving and processing
the raw data needed to make those variables, and returning them to you in either a
dataframe or list of dataframes. For the function to work, you need to be able
to direct it to the local directory that is home to the raw (i.e., as received)
MoBa phenotypic data files >=v11 in SPSS format. Remember, datasets can only be
curated on the basis of variables that exist in your raw MoBa data files. If something
is not working, it is worth checking MoBa documentation for item codes and verifying that
they exist in your raw data.

For a full introduction to the \code{curate_dataset} function, see the package
vignette (\code{vignette("phenotools")}).
}
\examples{

# Curate a scale- and item-level dataet from MoBa only, returned as a list

mydata <- curate_dataset(variables_required=c("scl_anx_m_3yr","bmi_derived_f_q1"),return_items = T, out_format="list")

# Curate a scale- and item-level dataet from MoBa only, returned as a merged dataframe

mydata <- curate_dataset(variables_required=c("scl_anx_m_3yr","bmi_derived_f_q1"),return_items = T, out_format="merged_df")

# Curate a dataset from multiple sources

mydata <- curate_dataset(variables_required=list(moba = c("scl_anx_m_3yr","bmi_derived_f_q1"),
npr = random5nprcodes),
return_items = T, out_format="merged_df",
exclusions=NULL,
recursive=TRUE,
group_all=TRUE,
dx_groupname="ourdiagnoses",
dx_owner="child")

For more examples, see the package vignette (\code{vignette("phenotools")}).
}
