% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_dataset.R
\name{curate_dataset}
\alias{curate_dataset}
\title{Initialise project}
\usage{
curate_dataset(variables_required = "none_specified",
  pheno_data_root_dir = "N:/data/durable/data/MoBaPhenoData/PDB2306_MoBa_V12/SPSS/",
  PDB = "2306", completion_threshold = 0.5, return_items = FALSE,
  consistent_items = FALSE, transformations = NULL)
}
\arguments{
\item{variables_required}{What variables are required? See
\code{available_variables} for a valid inputs}

\item{pheno_data_root_dir}{Where is the raw MoBa phenotypic data?}

\item{PDB}{What is the PDB code for your TSD project?}

\item{completion_threshold}{What proportion of scale items need
to be present for a scale score to be computed?}

\item{return_items}{Output item-level data?}

\item{consistent_items}{Only use wave-to-wave consistent items
for variables that are measured longitudinally?}

\item{transformations}{Not yet implemented}
}
\description{
\code{initialise_project} populates a specified directory
with the file structure required to make full use of phenotools
}
\details{
Detailed description...
}
