% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_npr.R
\name{curate_npr}
\alias{curate_npr}
\title{Curate NPR data}
\usage{
curate_npr(
  diagnoses,
  recursive = TRUE,
  exclusions = NULL,
  group_all = FALSE,
  dx_groupname = NULL,
  dx_recipient = c("child"),
  dx_range_limits = NULL,
  npr_full = NULL,
  npr_preprocessed = NULL,
 
    moba_data_root_dir = "//ess01/P471/data/durable/data/MoBaPhenoData/PDB2306_MoBa_V12/SPSS/",
  npr_data_root_dir = "//ess01/P471/data/durable/data/NPR/processed/",
  npr_filename = "npr2024.sav",
  npr_linkage_file_root_dir = "//ess01/P471/data/durable/data/Linkage_files/NPR_link/",
  npr_linkage_filename = "PDB2306_kobling_npr2024_mor_far_barn.sav",
  PDB = "2306",
  moba_data_version = 12,
  moba_filepaths = NULL,
  ...
)
}
\arguments{
\item{diagnoses}{ICD codes (see \code{available_variables(source=\"npr\")})}

\item{recursive}{also include sub-codes of codes you have provided in the
diagnoses field? Defaults to TRUE}

\item{exclusions}{codes to exclude - always non-recursive (i.e., only the
specific codes you provide will be excluded from individual-level summaries, not sub-codes)}

\item{group_all}{If TRUE, will treat all your codes as a single diagnosis
for the purpose of summarising; defaults to FALSE}

\item{dx_groupname}{If group_all is TRUE, this string is used to label the
vareiables pertaining to the grouped diagnoses}

\item{dx_recipient}{Whose diagnoses do you want to count? Dataset is returned as
one row per pregnancy, but you can get the diagnoses relevant to any
combination of "child", "father", and "mother"; the default is "child"}

\item{dx_range_limits}{If you want only to extract diagnoses from a particular
range of available years, provide limits here (e.g., "c(2008,2010)" will retrieve
diagnoses from 2008, 2009, and 2010 only). Defaults to NULL - i.e., extract
from all available data}

\item{npr_full}{if you already have the NPR data loaded in memory, you can
save some time by providing the name of the R object here; note
that there are specific requirements for the structure of this file (see
Details) below.}

\item{npr_preprocessed}{if you have already got a version of the NPR data
linked with MoBa IDs loaded in memory, supply the R object name here; note
that there are specific requirements for the structure of this file (see
Details) below.}

\item{moba_data_root_dir}{Where is the raw MoBa phenotypic data? (default
is for p471)}

\item{npr_data_root_dir}{Where is the raw NPR phenotypic data? (default
is for p471)}

\item{npr_filename}{What is the name of the NPR file? (default
is for p471)}

\item{npr_linkage_file_root_dir}{Where is the linkage file? (default
is for p471)}

\item{npr_linkage_filename}{What is the name of the linkage file? (default
is for p471). Required variables are: LNr (link number); M_ID_>PDB< (maternal ID);
F_ID_>PDB< (paternal ID); PREG_ID_>PDB< (preg ID); BARN_NR (child number in pregnancy);
SUP_Type (either MOTHER/FATHER/CHILD)}

\item{PDB}{What is the PDB code for your TSD project? (default
is for p471)}

\item{moba_data_version}{What version is the MoBa data you are linking to? Defaults to
12}

\item{moba_filepaths}{Allows curate_dataset to pass on an amended list of filepaths}
}
\description{
\code{curate_npr} is called by \code{curate_dataset}
when npr variables are requested - options for this function are included
directly in your \code{curate_dataset} call and passed on internally
}
\details{
This internal function is called when you list ICD-10 diagnostic codes in your
\code{curate_dataset} call. It's default is to work directly on a raw file
delivered by NPR, but since deliveries often vary in structure and content
in unpredictable ways, you may need to pre-process your NPR data for phenotools
to be able to work with it. Necessary formatting conditions are as follows:

- Data is in .sav format
- One row per healthcare interaction (individuals recur across rows)
- NPR ID column named either "LNR" or "LOPENR" (case insensitive)
- Diagnoses in columns prefixed with "tilst" (usually as delivered)
- Columns: innDato,utDato,henvType,omsorgsniva3,SUPtype (as delivered)

If your raw file meets these conditions you should be able to direct the function
to it \code{npr_filename} and \code{npr_data_root_dir} options in your
\code{curate_dataset}) call. If not, you should create a version of the file
meeting these conditions for use with phenotools.

The next step, prior to processing diagnoses, is to link with MoBa IDs. Again,
the function will do this automatically if the linkage file (specified via
 \code{npr_linkage_filename} and \code{npr_linkage_file_root_dir}) is in the standard
 format, i.e., containing linkage information for all members of family trios
 with the columns preg_id, m_id, f_id (all case insensitive), BARN_NR, LNr, and
 the variable SUPtype (case insensitive) - which is either "child", "mother",
 or "father" depending on who is the owner of the LNr. If the linkage file is
 not in this format, either format the file to match or pre-link the data yourself,
 in which case it should be read into memory and supplied using the option
  \code{npr_preprocessed}
}
