% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupsum.R
\name{dupsum}
\alias{dupsum}
\title{function to output a summary table of precision and recall}
\usage{
dupsum(true, test = list())
}
\arguments{
\item{true}{unique pairs of true duplicated characters}

\item{test}{list of pairs of potential duplicated characters (each member of list is a different method)}
}
\value{
a data frame object with false negatives (FN), true positives (TP), false positive (FP), along with precision (TP/(TP+FP)), recall (TP/(TP+FN)), and F1 score (2 * (precision * recall) / (precision + recall))
}
\description{
function to output a summary table of precision and recall
}
