% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupsum.R
\name{dupsum}
\alias{dupsum}
\title{function to output a summary table of precision and recall}
\usage{
dupsum(true, test = list())
}
\arguments{
\item{true}{unique pairs of true duplicated characters}

\item{test}{list of pairs of potential duplicated characters (each member of list is a different method)}
}
\description{
function to output a summary table of precision and recall
}
