% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcomments.R
\name{getcomments}
\alias{getcomments}
\title{Function to find comments in text file}
\usage{
getcomments(file)
}
\arguments{
\item{file}{name of file}
}
\description{
This function reads in a character list (as a text file) that have been annotated
as to whether a given character is a duplicate (DD), should be cut (XX), a
good character that should be kept (KK), or a character that should be kept but
modified (KKM).
}
