% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcomments.R
\name{getcomments}
\alias{getcomments}
\title{Function to find comments in text file}
\usage{
getcomments(file)
}
\arguments{
\item{file}{name of file}
}
\description{
How characters are characterized
DD = duplicate
KK = keep
KKM = keep and modify wording
XX = cut
}
