% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nex.R
\name{plot.nex}
\alias{plot.nex}
\title{plot data matrix and phylogeny}
\usage{
\method{plot}{nex}(x, phy = NULL, legend.pos = c("none", "left",
  "right", "bottom", "top"), bw = FALSE, na.value = "lightgray",
  fsize = 8, fill = c("statelabels", "charpartition", "charset",
  "file"))
}
\arguments{
\item{x}{a `nex` object for plotting}

\item{phy}{a phylogeny in `phylo` format}

\item{legend.pos}{position of legend}

\item{bw}{whether to plot in black (scoring present) and white (no scoring)}

\item{na.value}{color for NA values in matrix}

\item{fsize}{font size}

\item{fill}{how to color cells in matrix}
}
\description{
Plots a nexus dataset and optionally a tree
}
\examples{
\dontrun{
x <- allnex.genome
phy <- tree
}

}
