% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nex.R
\name{print.nex}
\alias{print.nex}
\title{Print data in a `nex` object}
\usage{
\method{print}{nex}(x, width = 6, ...)
}
\arguments{
\item{x}{nex class object}

\item{...}{other arguments (not used)}
}
\description{
Print data in a `nex` object
}
