% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nex.R
\name{read.nex}
\alias{read.nex}
\title{Read a nexus data file}
\usage{
read.nex(file, missing = "?", gap = "-")
}
\arguments{
\item{file}{(required) path to nexus file}

\item{missing}{character representing missing data}

\item{gap}{character representing inapplicable/incomporable data}
}
\value{
an object of class \code{nex} for use in further \code{nexustools} functions
}
\description{
A function that reads data stored in nexus files
}
\examples{
\dontrun{
x <- read.nex(file = system.file("extdata", "clarke_2006.nex", package = "phenotools"))
x
plot(x)
}

}
\author{
Chad Eliason \email{celiason@fieldmuseum.org}
}
