% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ontology.R
\name{read_ontology}
\alias{read_ontology}
\title{Function to create network graph from tab-based text file}
\usage{
read_ontology(file, root = FALSE)
}
\arguments{
\item{file}{path to text file containing trait ontology (see Details)}

\item{root}{whether to add a root node

e.g.,
feather
    rachis 
  barb
        barbule
            hooklet}
}
\description{
Function to create network graph from tab-based text file
}
