% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.nex.R
\name{sort.nex}
\alias{sort.nex}
\title{sort nexus file by something}
\usage{
\method{sort}{nex}(x, by = c("taxlabels", "charlabels"), ...)
}
\arguments{
\item{x}{nexus object}

\item{by}{sorting by}

\item{...}{other arguments passed to sort}
}
\description{
sort nexus file by something
}
