% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nex.R
\name{tail.nex}
\alias{tail.nex}
\title{Look at bottom of a `nex` object}
\usage{
\method{tail}{nex}(x, width = 6, ...)
}
\arguments{
\item{x}{`nex` class object}

\item{width}{width of character string to display}

\item{...}{optional arguments}
}
\description{
Look at bottom of a `nex` object
}
