% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlink.R
\name{traitlink}
\alias{traitlink}
\title{Function to link character descriptions to a trait ontology}
\usage{
traitlink(tree, x)
}
\arguments{
\item{tree}{trait ontology (igraph object)}

\item{x}{nexus file}
}
\value{
an object of class XX
}
\description{
Function uses fuzzy text matching to output a list of potentially redundant
characters in a `nex` object, waits for user input to confirm, and then
handles merging of characters and outputs a new `nex` object
}
\examples{
\dontrun{
x <- twig2.ss
tree <- tree
x=nex
}

}
\author{
Chad M. Eliason
}
