% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig-data.R
\docType{data}
\name{twig}
\alias{twig}
\title{Morphological character dataset for theropod dinosaurs}
\format{An object of class \code{"nex"}; see \code{read.nexus}.}
\usage{
data(twig)
}
\description{
Data assembled from assessment of study specimens
}
\examples{
data(twig)
plot(twig)
twig$taxlabels
}
\references{
Brusatte et al. (2014) Current Biology 24:2386-2392
(\href{https://www.ncbi.nlm.nih.gov/pubmed/25264248}{PubMed})
}
\keyword{datasets}
