% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.nex.R
\name{update.nex}
\alias{update.nex}
\title{Small function to update a duplicated nex object}
\usage{
\method{update}{nex}(x, cutoff = Inf)
}
\arguments{
\item{x}{nex object}

\item{cutoff}{cuttoff to filter duplicates/overlapping characters by}
}
\description{
changing cutoff will affect the characters identified as duplicates for later
visualization (e.g., with `printout` or `duptree` functions)
}
